requirements.txt
asyncpg~=0.25.0
python-dotenv~=0.20.0
aiohttp~=3.8.1
sqlalchemy~=1.4.36
PyYAML~=6.0
git+https://github.com/Rapptz/discord.py.git#egg=discord.py
